change OverflowProcessor.java for fix sonar
