+++ iotdb/src/main/java/org/apache/iotdb/db/engine/pool/MergeManager.java
@@ -116,7 +116,9 @@ public class MergeManager {
   }
 
   private static class InstanceHolder {
-
+    private InstanceHolder(){
+      //allowed do nothing
+    }
     private static MergeManager instance = new MergeManager();
   }
 }
