+++ iotdb/src/main/java/org/apache/iotdb/db/engine/querycontext/OverflowInsertFile.java
@@ -29,6 +29,7 @@ public class OverflowInsertFile {
   private List<ChunkMetaData> timeSeriesChunkMetaData;
 
   public OverflowInsertFile() {
+    //allowed do nothing
   }
 
   public OverflowInsertFile(String path, List<ChunkMetaData> timeSeriesChunkMetaData) {
