+++ iotdb/src/test/java/org/apache/iotdb/db/engine/overflow/iov2/OverflowProcessorBenchmark.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.engine.overflow.ioV2;
+package org.apache.iotdb.db.engine.overflow.iov2;
 
 import java.io.File;
 import java.io.IOException;
